requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.26.0
python-dotenv==1.0.1

langchain==0.0.353
langchain-core==0.0.23
langchain-openai==0.0.6
langchain-community==0.0.14

openai==1.16.2
supabase==2.0.0
tiktoken==0.6.0

dateparser